Issue number: Obtained from: Submitted by: Reviewed by:
This is an update that affects two services, createDataResource
and createContent. These were previously simple-method types
and now they are Java.

Note that these two services call the more complex "checkContentPermission"
service in order to determine permissions, but they have only been
checked out using a user with CONTENTMGR permission, so they only
demonstrate that the checkContentPermission can be substituted where the
simple entity permission check was used, since checkContentPermission
is a superset of the simpler check.


git-svn-id: http://svn.ofbiz.org/svn/ofbiz/trunk@656 deaff617-5be1-0310-a1b7-ec6878f92ba9