Issue number: Obtained from: Submitted by: Reviewed by:
A minor fix so that an "id" attribute will be added to fields that use
auto service/entity. A two line addition to mergeOverrideModelFormField
to handle idName ivar.

This was necessary in order to allow widget-form utilities to generate
textareas that will be compatible with the htmlarea widget.
The htmlarea widget requires an id attribute.


git-svn-id: http://svn.ofbiz.org/svn/ofbiz/trunk@680 deaff617-5be1-0310-a1b7-ec6878f92ba9