Issue number: Obtained from: Submitted by: Reviewed by:
I put some services and methods that deal with both Content and DataResource
at this level. Not happy about it, but not so sure where else to put them.
Worker has the "most recently used" code and Services adds a
persistContentAndAssoc service that is an improvement on another called
createContentAndAssoc.
The MapProcessors has some converters to take the prefixed names that come
from views and put them in std entity format. There are ones for
Content, DataResource and ContentAssoc.


git-svn-id: http://svn.ofbiz.org/svn/ofbiz/trunk@1114 deaff617-5be1-0310-a1b7-ec6878f92ba9