Issue number: Obtained from: Submitted by: Reviewed by:
I changed the htmlarea that this template was pointing to to the one
under content/webapp/content/images.
I had one under images/webapp/images and my CVS Entries showed it as part
of the CVS, but on my test server is did not show up in the checkout.
We probably talked about content being the proper place for htmlarea anyway.


git-svn-id: http://svn.ofbiz.org/svn/ofbiz/trunk@1155 deaff617-5be1-0310-a1b7-ec6878f92ba9