Issue number: Obtained from: Submitted by: Reviewed by:
LoopSubContentTransform is a new transform that take multiple subContent,
such as a blog log, and render them.
It allows embedded edit/RenderSubContent transforms.
I use an editRenderSubContent to attach the "View", "Respond" buttons.

The LoopWriter is necessary in order to include the "onStart" and "afterBody"
methods that make the subContent accessible.


git-svn-id: http://svn.ofbiz.org/svn/ofbiz/trunk@1354 deaff617-5be1-0310-a1b7-ec6878f92ba9