Issue number: Obtained from: Submitted by: Reviewed by:
This is a commit of changes to the Loop and Traverse transforms that make
them work and it is also a commit of the four existing transforms
(EditRender, Render, Loop and Traverse) in Cache mode.
No changes were made to the two transforms, EditRender and Render, that
would be being currently used.

The main change required for caching was the writing of a cached version of
content/ContentWorker.getAssocAndContentAndDataResource as a "Cache" version.
I still use the same views as before, it just builds them using cache methods
on the individual entity components of the view so that the cache maintenance
will still work.

In the cache transforms, I also cleaned up the interaction between the transforms
and wrote some utilities that are in FreeMarkerWorker (ie. saveValues, reloadValues)
for handling the exchange of data with the context object. Makes the code much
cleaner.

I will be working only with the cache transform versions for now. Once they are
stable, it will be trivial to convert the non-cache versions to take advantage
of the refinements in the cached versions.


git-svn-id: http://svn.ofbiz.org/svn/ofbiz/trunk@1446 deaff617-5be1-0310-a1b7-ec6878f92ba9