Issue number: Obtained from: Submitted by: Reviewed by:
Added InjectNodeTrailCsvTransform as a way of passing a navigation trail.
Other changes are fixes to the context save/restore scheme that came up
in testing.
Of course, I made changes to FreeMarkerWorker to tie in Inject...
Also some utilities in FreeMarkerWorker to convert a list of trail objects
to a concatenated string and back.


git-svn-id: http://svn.ofbiz.org/svn/ofbiz/trunk@1470 deaff617-5be1-0310-a1b7-ec6878f92ba9