Issue number: Obtained from: Submitted by: Reviewed by:
I can't seem to get the menu.xml def files to validate by giving it a
local URI at the top of the xml file. I have tried:
<!DOCTYPE menus PUBLIC "-//OFBiz//DTD OFBiz Form Widget//EN" "http://localhost:8443/ageps/images/widget-menu.dtd">
<!DOCTYPE menus PUBLIC "-//OFBiz//DTD OFBiz Form Widget//EN" "file:///usr/local/agi/ofbiz/hot-deploy/ageps/webapp/ageps/images/widget-menu.dtd">
<!DOCTYPE menus PUBLIC "-//OFBiz//DTD OFBiz Form Widget//EN" "http://www.ofbiz.org/dtds/widget-menu.dtd">

There is no error message that the dtd is not written correctly; all it says is
"name" must be declared...

So I thought if you can put widget-menu.dtd in the same place on www.ofbiz.org as
widget-form.dtd, then I can rule not finding it out.


git-svn-id: http://svn.ofbiz.org/svn/ofbiz/trunk@1877 deaff617-5be1-0310-a1b7-ec6878f92ba9