Issue number: Obtained from: Submitted by: Reviewed by:
This is a major commit of the changes that I have made to the Content module.
It is a little before I am ready, as there has been no documentation added,
but someone has had trouble getting the Content module to compile, so I thought
I would commit what I have. It has been working for some time.
As soon as this commit is done, I will commit some other auxillary files and the
do a new check out on another machine, as I am certaint that this will probably
break the build but I will have it working as soon as possible.

MenuWrapTranform is a new FreeMarker transform that allow the ModelMenu
stuff to be used at the FM level. It can be used to add "function bars"
to pages.

WrapSubContentCacheTransform separates the addition of an "Edit" button from the
EditSubContentTransform. My early decision to make this an option of Edit transform
compromised the design in several places. This basically allows any template to
"wrap" what is inside - an edit button being the most common application.

ContentManagementEvents adds two methods to work with Content. Those methods are not
now being used, but could be used in the future.

PermissionRecorder is an attempt to deal with the thorny issue of debugging permission
schemes. It records all the individual permission tests outputs them in an HTML
table with green and red colors used to indicate pass/fail. It is currently used
in ContentServices.createContentAssoc to display the table in the error screen if
it fails the permission check. In the future, there should be other places and ways
to make the information available.

In ContentManagementServices, I did some work to make the persistContentAndAssoc
more useful. I fixed some error handling among other things.

Added a lot of methods to ContentManagementWorker. Two (getCurrentValue...) are designed
to capture some code that is commonly used in JPublish beanshell action scripts, so that
it is not repeated so often.

Almost everything else in ContentMangementWorker are tools to facilitate the managment
of "publish points" and "content departments". These are content types that I set up
(WEB_SITE_PUBLISH_POINT and DEPARTMENT) to help in catagorizing and displaying content
and in setting up a permission scheme so that there will be places to assign persons so
that they have the right permissions for authoring, editing, admin and publishing.

Made a lot of changes to ContentPermissionServices to add more variables to the mix
of things that can be checked. In each case I kept the old API and added an overloaded
method with the added parameters, but that has built up 4 or 5 methods and I probably need
to go in and remove some now. Also added code to work with the PermissionRecorder.

In ContentServices and in other places I corrected the case where there was a mismatch of
target operation types. At one point I was using CREATE_CONTENT, UPDATE_CONTENT, etc, but
the ContentPurposeOperation table was using CONTENT_CREATE, CONTENT_UPDATE, etc. I switched
everything to match the ContentPurposeOperation table.

In ContentServices.xml, I added a fix to createContentRole to see if there role was already
attached to party and, if not, add that relationship so that referential integrity would exist.
Also removed a redundant method - createContentRole.

In ContentServicesComplex.java I added a parameter to the getAssocAndContentAndDataResource
to let it only find entities where the fromDate is null. This should cut out a filterByDate
operation and make display faster.

In ContentWorker.traverseSubContent I made some fixes to the way the beanshell logic for
pick and continuing was done. These are beanshell scripts that can be defined in FM
transforms and jpublish pagedef files to control how content structures are traversed
and displayed.

In ContentWorker, I also added checkConditions, which consolidated a lot of the beanshell
checking to see if the content should be included or not. I moved the permission checking
from the point of use to the time when the child structure is built in "selectKids()".

In UpdateContentAndImage I added some code to add an Author association. The addition of
an AUTHOR ContentType IS ONE OF THE BIG DEVIATIONS FROM WHAT HAD BEEN TALKED ABOUT. I was
relying on the createdBy contents to identity the author of content, but as I started
building my site, I realized that I would want to have a section for people and that I
would want to attach images and bio content to this content and use the facilities in
the content module for displaying these pieces of content, so it would be useful to make
people content instead of just handling them thru the Party and Person entities. I still
use Party, Person and UserLogin as they are currently used, I just have a Content entity
for each UserLogin that shares the userLoginId as its primary key.

In data/DataResourceWorker.renderDataResourceAsTextCache I implemented a cache for FM
template content as a speed enhancement. It can be controlled from the content.properties
config file.

In data/DataResourceWorker.writeDataResourceText there was a situation in which there
is mixed and confusing use of the variable, templateContext. In this case it is actually
a templateRoot (created by the FreeMarkerWorker.createEnvironmentMap()) and it contains
a context Map object which really contains the required values. For now, I put code
that looks for values in both places. This is the best example of something that needs
to be cleaned up.

Not much in DataServices except some CREATE_CONTENT -> CONTENT_CREATE fixes.

In webapp/ftl/CheckPermissionTransform there were fixes made.

EditSubContentTransform should be deprecated because of the addition of the
WrapSubContentCacheTransform.

Lots of changes in FreeMarkerWorker. The addition of the new transforms to
addAllOfbizTransforms() for one. Also, changes to how the node stack is passed
between transforms. This is the biggest difference between cached and non-cached
transforms (other than their use of caching) and probably makes their APIs
incompatible. It has been months since I have used non-cached transforms.

Lots of changes to LoopSubContentCacheTransform. It does not traverse
any content, just does an initial query, but it does evaluate each
selected node using the same code as the TraverseSubContentCachTransform
so that the node on the "globalNodeTrail" stack will be ready for testing
within the Traverse transform.

The Loop transform also includes the values for pagination (viewIndex, viewSize, etc.)

The Loop transform and all the transforms handle the way in which it saves and reloads
values it receives. At one point I saved everything on entry and then restored it before
leaving. Now the transforms save the values they need before calling sub transforms.
But I also found cases (right now only the globalNodeTrail) that needed to be put back
into the environment so that they could be shared by peers, so I save some values on
entry and restore them.


git-svn-id: http://svn.ofbiz.org/svn/ofbiz/trunk@1942 deaff617-5be1-0310-a1b7-ec6878f92ba9