Issue number: Obtained from: Submitted by: Reviewed by:
In ContentWorker.checkConditions fixed a dependency on having a full
ContentAssocDataResource view.

In UploadContentAndImage wrote SUB_CONTENT as image contentAssocTypeId
instead of DESCRIPTION.

In DataEvents.serveImage - added code to read an image from a file.

In DataResourceImage - added override code to acquireImage

In TraverseSubContentCacheTransform - added code to handle a non-existent
nodeTrail string.


git-svn-id: http://svn.ofbiz.org/svn/ofbiz/trunk@2005 deaff617-5be1-0310-a1b7-ec6878f92ba9