Issue number: Obtained from: Submitted by: Reviewed by:
ContentManagementServices: Added mapKey to persistContentAndAssoc method.
Added support for binary files.
Added test to better catch errors in called services.
Deactivated existing associations when adding a new one.

ContentPermissionServices: Realized that I wanted to be able to send in
Operation ids and purposes in string form so that it can be used directly
in events, so I added code that will handle both string and list forms of
those two parameters.
Moved the test for _OWNER_ so that it is not carried up to the parent content;
the _OWNER_ test is redone for each ancestor.

ContentServices and DataServices: Updated create and update methods to receive operation
and purpose parameters in string or list form per above.

ContentWorker: Broke up a callContentPermission method so that there is
also a version that returns the service result, as well as the original
which returned a string.
This is so the permissionRecorder object can be obtained, if desired.
Did something similar in DataResourceWorker.
Added prepTargetOperationList and prepContentPurposeList methods to
implement the dual input type handling mentioned above.

UploadContentAndImage: Modified uploadContentAndImage so that it
be used directly from different screens.


git-svn-id: http://svn.ofbiz.org/svn/ofbiz/trunk@2066 deaff617-5be1-0310-a1b7-ec6878f92ba9