Issue number: Obtained from: Submitted by: Reviewed by:
LoopSubContentCacheTransform: Fixed so that pagination works.

In the widget menu methods, code was added that would reinitialize menus
when the userLogin changed. A corresponding change was made to the ageps
controller.xml to call the event first, then the login scripts.

All other changes were the addition or deletion of debugging statements.


git-svn-id: http://svn.ofbiz.org/svn/ofbiz/trunk@2073 deaff617-5be1-0310-a1b7-ec6878f92ba9