Issue number: Obtained from: Submitted by: Reviewed by:
ContentManagementServices: in persistContentAndAssoc, added the flag
"deactivateExisting" to control whether existing ContentAssocs are deactivated.

ContentManagementWorker: fixed a bug in "getFromSomewhere" to handle the
case where the value is not a String object.

FreeMarkerWorker: in saveContextValues and reloadValues, did a better job
of "cloning" lists and maps so that they don't get clobbered below.

Loop/TraverseSubContentCacheTransform: major reorganization to control
how the bread crumb trail is handled. The combination of faulty saving/restoring
and cluttered use of the "globalNodeTrail" var made things a mess.


git-svn-id: http://svn.ofbiz.org/svn/ofbiz/trunk@2075 deaff617-5be1-0310-a1b7-ec6878f92ba9