Issue number: Obtained from: Submitted by: Reviewed by:
I took out all the logVerbose commands that I had put in.
I could not use them because when I turned verboseOn, it goes into a
nested condition - I believe it has to do with a value that I am putting
into a context store, but I had not used any of the logVerbose commands
for sometime and had added plenty of logInfo, so I felt it was not
hurting anything to take them out.


git-svn-id: http://svn.ofbiz.org/svn/ofbiz/trunk@2111 deaff617-5be1-0310-a1b7-ec6878f92ba9