Issue number: Obtained from: Submitted by: Reviewed by:
These are image content and javascript files that were in components/images
(and I did not remove them from there) that I needed in components/content
because I was having trouble with resolution when going thru the proxy process.
I had images that were reference as "/images/cal.gif" in content forms, but
when I had automationgroups.com handled as a virtual host with a mount point
of "/" then it conflicts with ofbiz components/images.


git-svn-id: http://svn.ofbiz.org/svn/ofbiz/trunk@2168 deaff617-5be1-0310-a1b7-ec6878f92ba9