Issue number: Obtained from: Submitted by: Reviewed by:
These mods are related to refactoring to make the administration of
content management more workable.

RenderSub/ContentAsText are transforms that are meant to replace
RenderSubContentCacheTransform. The follow the ContentWorker.renderSub/ContentAsText
methods a little more closely. RenderSubContentCacheTransform accepts
contentId and mapKey, and subContentId as separate arguments.

Many other changes were put in place to change how images and binary data are
handled. Now, storing and retrieving images from binary files, in addition to
the ImageDataResource entity, are handled.


git-svn-id: http://svn.ofbiz.org/svn/ofbiz/trunk@2749 deaff617-5be1-0310-a1b7-ec6878f92ba9