Issue number: Obtained from: Submitted by: Reviewed by:
Other code to go along with tree widget source.
In ModelScreenWidget, I modified the Tree.renderWidgetString method to
look at the "render-class" attribute of the tree xml element to see
which class to instantiate as the TreeStringRenderer class.
That way one can program in the tree config file to run
HtmlTreeRenderer or HtmlTreeExpandCollapseRenderer.


git-svn-id: http://svn.ofbiz.org/svn/ofbiz/trunk@2868 deaff617-5be1-0310-a1b7-ec6878f92ba9