Issue number: Obtained from: Submitted by: Reviewed by:
I put ModelScreenWidget back so that it knows which class to implement
for each node render-style.

I also added the same inner classes, Link and Image, to ModelScreenWidget
that I added to ModelTree - with some minor changes in order to integerate.
I thought I got the go ahead for Link and I am just assuming that Image
is needed. The only other way I see to add it is via content, which
seems like overkill. Anyway, I figured it would be easy enough to back it out.

I changed HtmlTreeExpandCollapseRenderer to use images instead of text +/- s.


git-svn-id: http://svn.ofbiz.org/svn/ofbiz/trunk@2884 deaff617-5be1-0310-a1b7-ec6878f92ba9