Issue number: Obtained from: Submitted by: Reviewed by:
Changes related to 2 things:

1. In ModelTree i tested the screen component for the first time.
I made some changes in ModelScreen.ScreenRenderer to get the ScreenStringRenderer
so that I could use it in ModelTree. If it does not exist, then a
new ScreenStringRenderer is created in HtmlTreeRenderer.

2. Refactored ModelMenu items to use a Link inner class like ModelScreen/Tree.
I left the old code in, but i would like to take it out.

Changes to HtmlScreenRenderer and ModelScreenWidget was just a debug line that
i put in and then took out.


git-svn-id: http://svn.ofbiz.org/svn/ofbiz/trunk@2893 deaff617-5be1-0310-a1b7-ec6878f92ba9