Many changes related to making content/cms screens work with widget-screen technology.
In widget-form an actions element was added like in widget-screen.
A ModelFormAction class was added to support the actions.
The EntityAnd, EntityCondition and Service subclasses were altered to pass an
EntityListIterator object in the context.

ModelMenu was changed to get rid of MenuTarget and related classes. 
A lot of lines were removed from widget-menu.




git-svn-id: http://svn.ofbiz.org/svn/ofbiz/trunk@3158 deaff617-5be1-0310-a1b7-ec6878f92ba9