Removed old ModelTree in widgets/screen
Changed ModelScreenWidget.Sub/Content objects to take "edit-request" attribute to
indicate what request to use to edit content data.

Changed the way that ModelTree is called. Added an expandRequest name, but it
defaults to the current request name. Added a "trail" name attribute. It
defaults to "trail". So neither attribute is required.

Changed ModelTree so that its entity-one does not use "value-name".

Changed ModelTree and HtmlTreeExpandCollapseRenderer so that the "breadcrumb" trail
is kept in ModelTree so it can be used by any rendere.



git-svn-id: http://svn.ofbiz.org/svn/ofbiz/trunk@3268 deaff617-5be1-0310-a1b7-ec6878f92ba9