Upgraded ContentManagementServices.persistContentAndAssoc so that it would use multipart upload parameters (_imageData_contentType and _imageData_fileName).
I also made sure that persistContentAndAssoc returns all keys to Content, DataResource
and ContentAssoc passed to it so that it can be used directly in more situations
to populate result forms.

Made a significant change in the way that widget trees are rendered.
I did have two render classes, HtmlTreeRenderer and HtmlExpandCollapseRenderer,
but I eliminated HtmlExpandCollapseRenderer and made the expand-collapse just
a feature that is turned off and on by the tree element's "render-style".
I also added new styles - "follow-trail" (for situations like a bread crumb trail
where you only want a path rendered) and "show-peer" which would be like
expand-collapse - showing peers along the trail.



git-svn-id: http://svn.ofbiz.org/svn/ofbiz/trunk@3565 deaff617-5be1-0310-a1b7-ec6878f92ba9