Added "entry-name" and "list-name" to ModelTree so that it would be consistent with the ModelScreenWidget.IterateTree class. It turns out that "list-name" is not really necessary, but it is optional, as is "entry-name". Doing so allows ModelTree and IterateSection code to call the same "include-screen". Without doing so, the string expanders would have a sub map qualifier in one case and not in another.
I also changed the signature of renderNodeActionBegin, but since it is only
used within ModelTree, I can't see that it is a big deal. 
I needed to fix the way in which a tree determines if it has children,
so I needed to save the subNodeValues between calls to hasChildren
and getChildren and making subNodeValues an instance var seemed to be the best way.


git-svn-id: http://svn.ofbiz.org/svn/ofbiz/trunk@3951 deaff617-5be1-0310-a1b7-ec6878f92ba9