I upgrade EntityPermissionChecker to use three inner classes of ContentPermissionServices, PermissionConditionGetter, RelatedRoleGetter and AuxiliaryValueGetter (for ContentPurposes). Though I had coded ContentPermissionServices.checkPermissionMethod with the idea of having it work for non-Content permission checking, it would not do it, so I added another checkPermissionMethod using the three classes above.
This does not break any widget forms. If none of the three subelements of if-entity-permission
are defined, it will use defaults that make it work like it did before.

With this upgrade I should be able to work with PartyRelationship and WorkEffortAndPartyAssign
to create substantial websites using use the widget framework.


git-svn-id: http://svn.ofbiz.org/svn/ofbiz/trunk@4029 deaff617-5be1-0310-a1b7-ec6878f92ba9