All the form widget stuff has to do with adding "field-group" and "banner" tags to the form widget schema so that the forms can be broken up. Also added "skip-start" and "skip-end" attributes to "form" so that multiple form includes can be called to build one form. This would allow freemarker templates to be mixed with the form def.
I modified the renderContent/SubContent/DataResourceAsText services
(but not their underlying method calls, so speed should not be affected)
to return the string that is written to the writer.
This will allow those services to be used in screen widget def files.

I did a lot of work on permissions, filling in gaps where things had not
been used or tested.



git-svn-id: http://svn.ofbiz.org/svn/ofbiz/trunk@4133 deaff617-5be1-0310-a1b7-ec6878f92ba9