Implementation of a part of the requirements to purchase order process.
Mods to shopping cart item to store the associated requirement, if any.
During checkout the status of the requirements will be updated (from 'approved' to 'ordered')
Iif a requirement is directly linked to a sales order then the sales order and the purchase order lines are linked together thru the OrderItemAssociation entity.

Jira issue #OFBIZ-91.


git-svn-id: http://svn.ofbiz.org/svn/ofbiz/trunk@4484 deaff617-5be1-0310-a1b7-ec6878f92ba9