I've seen Si Chen's comment in the ftl template: "TODO: figure out how to share these files from the /includes directory"
So I've created a new screen definition for the company logo; then I've included the screen using the ${screens.rended("screenName")} directive.

This should be the right way to go.
Si, I've not moved the companyLogo.ftl.fo file to the includes folder: however, now you can do it if you think it's better (just remember to update the CompanyLogo screen definition).


git-svn-id: http://svn.ofbiz.org/svn/ofbiz/trunk@5402 deaff617-5be1-0310-a1b7-ec6878f92ba9