Addition of two services to derive Survey and SurveyResponses from PDF/FDF (AcroForm) files. They are still quite rough, but basically work.
There is currently no provision to related SurveyResponses with different field names.
For instance, the dd1423.pdf form that was used as a standard has some pretty weird fieldnames
and the Survey that was derived from it is not intuitive.
It is the SurveyQuestion.description field that has the non-intuitive info. 
The question field cannot be programmably set from a PDF/FDF file.
So, the question field could be filled with the appropriate question, it it would not 
matter what is in the description field - would it? I don't know what the description field is used for.



git-svn-id: http://svn.ofbiz.org/svn/ofbiz/trunk@6339 deaff617-5be1-0310-a1b7-ec6878f92ba9