A lot of changes to support CompDoc, but nothing much to demo.
Added some views to support find that most recent ContentRevision.

A number of screens, forms and trees to handle the display and editing of CompDocs.

Some seed data in CompDocData.xml.

Some services in CompDocServices.xml to assist with ContentApproval.

Added menus to CompDocMenus, including some to use for tree rendering.




git-svn-id: http://svn.ofbiz.org/svn/ofbiz/trunk@6606 deaff617-5be1-0310-a1b7-ec6878f92ba9