Added two new WorkEffort services: * setWorkEffortFixedAssetAssign * getProductManufacturingSummaryByFacility
The latter is useful to get information about the production runs (by facility) related to a given productId:
information about total quantity in production or total quantity needed by production runs is retrieved (together with the list of production runs).

This service is called by the bsh that prepares the data for the Product->Inventory page.

Also, the page has been improved:

* by default, the ATP/QOH/in&out shipments/in&out production runs are now shown only for the facilities to which the product is associated by the ProductFacility entity (and there is a link to show all the facilities)
* added information about total qty in/out from/to production runs



git-svn-id: http://svn.ofbiz.org/svn/ofbiz/trunk@7057 deaff617-5be1-0310-a1b7-ec6878f92ba9