This fixes some problems with earlier OpenOffice code not working with Microsoft Windows and it had to deal with problems caused by the latest UNO libraries.
I used the jar files from version 2.0.2. I am not sure if they are different from the Windows download to the Linux. The jar that I am currently using a ridl, unoil, java_uno_accesbridge, java_uno, juh, jurt, jut. I am not sure all of those are required. Eclipse says that only ridl and unoil are needed, but I thought there might be some future or indirect need for the others.

I am doing this work on a Windows laptop with .5G of memory, so I cannot adequately test. Will have to do that later.

git-svn-id: http://svn.ofbiz.org/svn/ofbiz/trunk@7428 deaff617-5be1-0310-a1b7-ec6878f92ba9