Applied patch from JIRA issue OFBIZ-1659. Following things are covered in this work :-
Restructured and organized the Screenc and Forms in the HR component.
Improved and Updated the naming conventions in the naming of Screens,Forms and requests. 

Thanks Harsha Chadhar for this work.
It was really very important for us to Manage the files before the component is used Extensively with other components.


git-svn-id: https://svn.apache.org/repos/asf/ofbiz/trunk@660980 13f79535-47bb-0310-9956-ffa450edef68