As suggested in dev ML: Currently the 'historic" promotions (PPIP_ORST_HIST,  PPIP_ORST_YEAR, PPIP_ORST_LAST_YEAR) do not include taxes in their calculation which uses remainingSubTotal. But, for B2C, in countries which are used to (or must) show prices with taxes (most of countries using VAT I guess), the price the client effectively pay is taxes included (the final client will never get back taxes paid, contrary to B2B case). So I propose to introduce a new field in ProductStore, namely includeTaxInRemainingSubTotal. If this field is set to Yes, then in the calculation of remainingSubTotal (done in service resetGrandTotal), the taxes will be included.
Actually my question is : in order to minize the number of fields and keep them consistent maybe we could simply use 
showPricesWithVatTax. It could be argued though that we really need to create the new field includeTaxInRemainingSubTotal in order 
to be able to deal with B2B case. But maybe it's no really needed because we set the field showPricesWithVatTax to Yes for the same 
basic reason that bring us to include taxes in remainingSubTotal calculation.

git-svn-id: https://svn.apache.org/repos/asf/ofbiz/trunk@744902 13f79535-47bb-0310-9956-ffa450edef68