Done following fixes: 1) Moved the view entity to correct package. 2) Changed the view name as well as the entities used. Now the records are coming as expected. 3) Corresponding changes in the findOrdersToPickMove service.
The main purpose of this view entity is to find out the location (areaId) of the inventory item that is reserved/picked for a particular order. For example if an order has two 
order items X and Y and items X is picked from area A and item Y is picked from area B of a facility (where obviously area A and B are associated to a pick/primary location), 
then while grouping this order in facility > picking screen by "Group by Warehouse Area" method will show order under Multi Locations group. To clarify more on this grouping
method, grouping is done only with the areaId and the locationSeqId (which is a combination of areaId, aisleId, sectionId, levelId, positionId).  


git-svn-id: https://svn.apache.org/repos/asf/ofbiz/trunk@779041 13f79535-47bb-0310-9956-ffa450edef68