GL Account Name (to the right of the "GL Account Id" field). Total of Current Fiscal Period (to the right of the "Total Amount" field): this is computed considering all the transactions posted to the account from the beginning of the Fiscal Period to the thruDate set by the user.
Added to the search for at the top of the screen an optional "Month" drop down field: it contains all the 12 months of the year; if a value (e.g. January) in this drop down is selected and the from/thru date are left empty then the system will compute the from/thru dates for the report considering the beginning and end of the selected month.


git-svn-id: https://svn.apache.org/repos/asf/ofbiz/trunk@795910 13f79535-47bb-0310-9956-ffa450edef68