Applied patch from jira issue OFBIZ-3167 - Add Constraint with respect to eBay root category for product search and change demo data according to the concept.
As we have different productCatalogCategoryType for eBay so add a constraint on searching that if catalog not associated with any PCCT_BROWSE_ROOT type category then also search for PCCT_EBAY_ROOT type category for CatalogConstrain. This will help to manage separate catalog and catalog category association for ebay.

Thanks Parimal for the contribution.

git-svn-id: https://svn.apache.org/repos/asf/ofbiz/trunk@833054 13f79535-47bb-0310-9956-ffa450edef68