Applied patch from jira issue OFBIZ-3508 -
When we create a new customer profile, it doesn't send correct mail according to "UPD_PRSNL_INF_CNFRM", "PARTY_REGIS_CONFIRM" and "RDS_CUST_REGISTER" emailType. Currently it grabs the first from the list of ProductStoreEmailSetting entity and uses it's email settings to send the email.
It was due to unavailability of productStoreId for sending mail so whenever new customer profile will be created it should be associated to certain product store and its role type will be "CUSTOMER".

Apart from this customer profile page should have information about associated product stores.

Following points are covered:
   1.  Improved patch from OFBIZ-3000.
   2. Improved Product Store Roles screens.
   3. Added ProductStores portlet into party profile screen.
   4. Created updateProductStoreRole service for updating ProductStoreRole record.
   5. Removed EditProductStoreRoles.groovy and EditProductStoreRoles.ftl .

Thanks to Amit Sharma for the contribution.

git-svn-id: https://svn.apache.org/repos/asf/ofbiz/trunk@916219 13f79535-47bb-0310-9956-ffa450edef68