Fixed a bug in vendor return. The service getReturnableItems returns a map containing the returable quantity corresponding to an item along with certain other things. Earlier this service was returning an empty map for a purchase order and a vendor return. On order detail page action block there is a check on this map and if the map comes out to be empty the system does not show the options to create return for the purchase order. Similary if you create a vendor return directly from the Return Header screen and try to add items on the Return Item screen using an order Id, this display a message "No returnable items found for order"
Also the method sendItemsShipped in the service sendGoogleShipRequest was throwing a NPE because the code try to pull a product through
ItemIssuance > OrderItem > Product

In the vendor return of type refund the items are issued to the return shipment and a purchase return invoice is generated and hence the process
is not linked to any Order. Adding a condition on the ECA so that this service is not invoked.



git-svn-id: https://svn.apache.org/repos/asf/ofbiz/trunk@940851 13f79535-47bb-0310-9956-ffa450edef68