A patch from BJ Freeman "complete web.xml for error reporting" (https://issues.apache.org/jira/browse/OFBIZ-3820) - OFBIZ-3820
BJ: when there is an error, the control servlet looks for an error path in <param-name>allowedPaths</param-name>
it does not find one throwing an exception which buries the orginal error.
instead of changing the code it is just easier to add the /error: path.
then the real error gets reported correctly.

JLR: This will maybe help to better understand some errors on demo servers, and can't hurt

git-svn-id: https://svn.apache.org/repos/asf/ofbiz/trunk@954135 13f79535-47bb-0310-9956-ffa450edef68